requirements.txt
flask
requests
gunicorn
python-decouple
flask-cors
celery==4.4.7
beautifulsoup4==4.8.1
redis
selenium
2captcha-python
flower